+++ test/services/concrete_object_service_test.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+require_relative '../test_helper'
+require 'services/concrete_object_service'
+class ConcreteObjectServiceTest < Minitest::Test
+  def test_calculates_area_and_volume_metadata
+    attributes = ForgeBuild::Services::ConcreteObjectService.new.attributes('slab_on_grade', 120, 240, 6)
+    assert_equal 28_800.0, attributes[:quantities][:area_sq_in]
+    assert_equal 172_800.0, attributes[:quantities][:volume_cu_in]
+    assert_equal '03', attributes[:csi_division]
+  end
+end